Make PHPStan happy (#36)
* Silence strict comparison using !== between null warnings
* Silence offset float|int might not exist on string warnings

+++ src/Base32/GmpEncoder.php
@@ -60,7 +60,7 @@ class GmpEncoder extends BaseEncoder
 
         /* Convert each five bits to a Base32 character. */
         $encoded = implode("", array_map(function ($fivebits) {
-            $index = bindec($fivebits);
+            $index = (int)bindec($fivebits);
             return $this->characters()[$index];
         }, $binary));
 
@@ -142,7 +142,7 @@ class GmpEncoder extends BaseEncoder
 
         /* Convert each five bits to a Base32 character. */
         $encoded = implode("", array_map(function ($fivebits) {
-            $index = bindec($fivebits);
+            $index = (int)bindec($fivebits);
             return $this->characters()[$index];
         }, $binary));
